stockmind-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
groq==0.4.1
httpx==0.25.1
aioredis==2.0.1
redis==5.0.1
yfinance==0.2.32
newsapi==0.1.1
langchain==0.1.0
langgraph==0.0.20
asyncio-throttle==1.0.2
python-multipart==0.0.6
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
bcrypt==4.1.1
supabase==2.4.5
python-jose==3.3.0
